application/requirements.txt
fastapi
uvicorn
pydantic
ultralytics
opencv-python
numpy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
sqlalchemy
